Show Line Id in the MainFrame
Currently the the line id / subsystem was not set as a text in the mainFrame.
Also the mechanism to download remotely the subsystem id was persisting the information in a separate global.properties files.
A future refactoring is needed in order to simplify the process, and propagate correctly the value form global.properties to the main frame.

See also KESCL-39